+++ R/deployApp.R
@@ -94,7 +94,9 @@
 #'   will always update the previously-deployed app. If `FALSE`, will ask
 #'   the user what to do, or fail if not in an interactive context.
 #'
-#'   Defaults to the value of `getOption("rsconnect.force.update.apps", FALSE)`.
+#'   Defaults to `TRUE` when called automatically by the IDE, and `FALSE`
+#'   otherwise. You can override the default by setting option
+#'   `rsconnect.force.update.apps`.
 #' @param python Full path to a python binary for use by `reticulate`.
 #'   Required if `reticulate` is a dependency of the app being deployed.
 #'   If python = NULL, and RETICULATE_PYTHON or RETICULATE_PYTHON_FALLBACK is
@@ -161,7 +163,7 @@ deployApp <- function(appDir = getwd(),
                       logLevel = c("normal", "quiet", "verbose"),
                       lint = TRUE,
                       metadata = list(),
-                      forceUpdate = getOption("rsconnect.force.update.apps", FALSE),
+                      forceUpdate = NULL,
                       python = NULL,
                       forceGeneratePythonEnvironment = FALSE,
                       quarto = NULL,
@@ -274,6 +276,9 @@ deployApp <- function(appDir = getwd(),
       server = server
     )
   } else {
+    forceUpdate <- forceUpdate %||% getOption("rsconnect.force.update.apps") %||%
+      fromIDE()
+
     # Use name/account/server to look up existing deployment;
     # create new deployment if no match found
     target <- deploymentTarget(
@@ -287,10 +292,10 @@ deployApp <- function(appDir = getwd(),
     )
   }
   if (is.null(target$appId)) {
-    dest <- accountId(target$username, target$server)
+    dest <- accountLabel(target$username, target$server)
     taskComplete(quiet, "Deploying {.val {target$appName}} to {.val {dest}}")
   } else {
-    dest <- accountId(target$username, target$server)
+    dest <- accountLabel(target$username, target$server)
     taskComplete(quiet, "Re-deploying {.val {target$appName}} to {.val {dest}}")
   }
 
@@ -337,7 +342,7 @@ deployApp <- function(appDir = getwd(),
       }
     )
     if (application$id == target$appId) {
-      taskComplete(quiet, "Found application")
+      taskComplete(quiet, "Found application {.url {application$url}}")
     } else {
       taskComplete(quiet, "Created application with id {.val {application$id}}")
     }
@@ -560,7 +565,7 @@ bundleApp <- function(appName,
     verbose = verbose,
     quiet = quiet
   )
-  manifestJson <- enc2utf8(toJSON(manifest, pretty = TRUE))
+  manifestJson <- toJSON(manifest)
   manifestPath <- file.path(bundleDir, "manifest.json")
   writeLines(manifestJson, manifestPath, useBytes = TRUE)