+++ tests/testthat/helper-http.R
@@ -22,6 +22,15 @@ local_cookie_store <- function(env = caller_env()) {
   withr::defer(env_bind(.cookieStore, !!!old), envir = env)
 }
 
+skip_on_http_failure <- function(code) {
+  tryCatch(
+    code,
+    rsconnect_http = function(cnd) {
+      testthat::skip("http request failed")
+    }
+  )
+}
+
 # Generic tests of various http methods -----------------------------------
 
 test_http_GET <- function() {
@@ -29,8 +38,7 @@ test_http_GET <- function() {
 
   # Perform the request
   resp <- GET(service, authInfo = NULL, path = "get")
-  expect_equal(attr(resp, "httpResponse")$status, 200)
-  expect_equal(attr(resp, "httpResponse")$contentType, "application/json")
+  expect_equal(attr(resp, "httpContentType"), "application/json")
   expect_equal(resp$path, "/get")
 }
 
@@ -39,7 +47,6 @@ test_http_POST_JSON <- function() {
 
   body <- list(a = 1, b = 2, c = 3)
   resp <- POST_JSON(service, authInfo = NULL, path = "post", json = body)
-  expect_equal(attr(resp, "httpResponse")$status, 200)
   expect_equal(resp$json, body)
 }
 
@@ -47,7 +54,6 @@ test_http_POST_empty <- function() {
   service <- httpbin_service()
 
   resp <- POST(service, authInfo = NULL, path = "post")
-  expect_equal(attr(resp, "httpResponse")$status, 200)
   expect_equal(resp$json, set_names(list()))
 }
 
@@ -66,7 +72,6 @@ test_http_POST_file <- function() {
     contentType = "text/plain",
     file = path
   )
-  expect_equal(attr(resp, "httpResponse")$status, 200)
   expect_equal(resp$data, "1\n2\n3\n")
 }
 
@@ -74,10 +79,8 @@ test_http_headers <- function() {
   service <- httpbin_service()
 
   resp <- GET(service, authInfo = list(apiKey = "abc123"), path = "get")
-  expect_equal(attr(resp, "httpResponse")$status, 200)
   expect_equal(resp$headers$Authorization, "Key abc123")
 
   resp <- POST(service, authInfo = list(apiKey = "abc123"), path = "post")
-  expect_equal(attr(resp, "httpResponse")$status, 200)
   expect_equal(resp$headers$Authorization, "Key abc123")
 }