+++ tests/testthat/test-http-libcurl.R
@@ -31,18 +31,18 @@ test_that("can get and set cookies", {
   service <- parseHttpUrl("http://httpbin.org/")
 
   # Setting
-  GET(service, list(), "cookies/set", query = "a=1")
+  skip_on_http_failure(GET(service, list(), "cookies/set", query = "a=1"))
   cookies <- getCookies(service$host)
   expect_equal(cookies$name, "a")
   expect_equal(cookies$value, "1")
 
   # Overwriting
-  GET(service, list(), "cookies/set", query = "a=2&b=1")
+  skip_on_http_failure(GET(service, list(), "cookies/set", query = "a=2&b=1"))
   cookies <- getCookies(service$host)
   expect_equal(cookies$name, c("b", "a"))
   expect_equal(cookies$value, c("1", "2"))
 
   # Preserved across calls
-  out <- GET(service, list(), "cookies")
+  skip_on_http_failure(out <- GET(service, list(), "cookies"))
   expect_equal(out$cookies, list(a = "2", b = "1"))
 })