+++ NEWS.md
@@ -10,6 +10,9 @@
   please file an issue then set `options(rsconnect.packrat = TRUE)` to revert 
   to the previous behaviour.
 
+* The default server name created by `addServer()` now includes the port,
+  if used.
+
 * New `rsconnect.http.headers` and `rsconnect.http.cookies` allow you to
   set extra arbitrary additional headers/cookies on each request (#405).
   Their use is documented in the new `vignette("custom-http")`.